+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -86,6 +86,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added a new mode that improves the filtering of ray traced shadows (directional, point and spot) based on the distance to the occluder.
 - Added support of cookie baking and add support on Disc light.
 - Added support for fog attenuation in path tracing.
+- Added a new debug panel for volumes
 
 ### Fixed
 - Update documentation of HDRISky-Backplate, precise how to have Ambient Occlusion on the Backplate